resultados/primera_vuelta/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
geopandas==1.1.3
idna==3.11
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pyogrio==0.12.1
pyproj==3.7.2
python-calamine==0.6.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.5
shapely==2.1.2
six==1.17.0
text-unidecode==1.3
urllib3==2.6.3
